requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.2.0
django-couchbase==0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
gunicorn==20.1.0